pyproject.toml
[project]
name = "rubick"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "lxml-html-clean>=0.4.3",
    "newspaper4k>=0.9.4.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "xlsxwriter>=3.2.9",
]
